[fixes #383] ensure every Exception inherits from KazooException
