+++ kazoo/retry.py
@@ -8,24 +8,13 @@ from kazoo.exceptions import (
     KazooException,
     OperationTimeoutError,
     SessionExpiredError,
+    ForceRetryError,
+    InterruptedError,
+    RetryFailedError
 )
 
+
 log = logging.getLogger(__name__)
-
-
-class ForceRetryError(Exception):
-    """Raised when some recipe logic wants to force a retry."""
-
-
-class RetryFailedError(KazooException):
-    """Raised when retrying an operation ultimately failed, after
-    retrying the maximum number of attempts.
-    """
-
-
-class InterruptedError(RetryFailedError):
-    """Raised when the retry is forcibly interrupted by the interrupt
-    function"""
 
 
 class KazooRetry(object):
